+++ src/gl/TextureAtlas.h
@@ -1,10 +1,11 @@
 #pragma once
+#include <stdint.h>
 
 class TextureNode
 {
 public:
 	TextureNode* child[2];
-	short x, y, w, h;
+	uint16_t x, y, w, h;
 	bool filled;
 
 	TextureNode(int offX, int offY, int mapW, int mapH);
@@ -12,7 +13,7 @@ public:
 
 	// places lightmap into the atlas, populating x/y coordinates
 	// info width/height must be set before calling
-	bool insert(int iw, int ih, int& outX, int& outY);
+	bool insert(int iw, int ih, uint16_t& outX, uint16_t& outY);
 };
 
 class TextureAtlas
@@ -32,5 +33,5 @@ public:
 	// places lightmap into the atlas, populating x/y coordinates
 	// info width/height must be set before calling
 	// id is used to select a sub-atlas (use face idex)
-	bool insert(int id, int iw, int ih, int& outX, int& outY);
+	bool insert(int id, int iw, int ih, uint16_t& outX, uint16_t& outY);
 };
